+++ openmls/src/framing/plaintext.rs
@@ -258,7 +258,7 @@ impl MlsPlaintext {
     }
 
     /// Remove the membership tag for testing.
-    #[cfg(any(feature = "expose-test-vectors", test))]
+    #[cfg(any(feature = "test-utils", test))]
     pub(crate) fn remove_membership_tag(&mut self) {
         self.membership_tag = None;
     }
