+++ openmls/src/group/tests/mod.rs
@@ -1,6 +1,6 @@
 //! Unit tests for the MLS group
 
-#[cfg(any(feature = "expose-test-vectors", test))]
+#[cfg(any(feature = "test-utils", test))]
 pub mod kat_messages;
-#[cfg(any(feature = "expose-test-vectors", test))]
+#[cfg(any(feature = "test-utils", test))]
 pub mod kat_transcripts;
