+++ openmls/src/lib.rs
@@ -27,9 +27,9 @@ mod utils;
 #[macro_use]
 pub mod error;
 
-#[cfg(any(feature = "expose-test-vectors", test))]
+#[cfg(any(feature = "test-utils", test))]
 #[macro_use]
-pub mod test_util;
+pub mod test_utils;
 
 pub mod ciphersuite;
 pub mod config;
@@ -40,9 +40,9 @@ pub mod group;
 mod key_packages;
 pub mod key_store;
 pub mod messages;
-#[cfg(any(feature = "expose-test-vectors", test))]
+#[cfg(any(feature = "test-utils", test))]
 pub mod schedule;
-#[cfg(not(any(feature = "expose-test-vectors", test)))]
+#[cfg(not(any(feature = "test-utils", test)))]
 mod schedule;
 pub mod tree;
