+++ openmls/src/schedule/kat_key_schedule.rs
@@ -12,12 +12,12 @@ use crate::{
     config::{Config, ProtocolVersion},
     group::{GroupContext, GroupEpoch, GroupId},
     schedule::{EpochSecrets, InitSecret, JoinerSecret, KeySchedule, WelcomeSecret},
-    test_util::{bytes_to_hex, hex_to_bytes},
+    test_utils::{bytes_to_hex, hex_to_bytes},
     utils::randombytes,
 };
 
 #[cfg(test)]
-use crate::test_util::{read, write};
+use crate::test_utils::{read, write};
 
 use hpke::HpkeKeyPair;
 use serde::{self, Deserialize, Serialize};
@@ -115,7 +115,7 @@ fn generate(
     )
 }
 
-#[cfg(any(feature = "expose-test-vectors", test))]
+#[cfg(any(feature = "test-utils", test))]
 pub fn generate_test_vector(
     n_epochs: u64,
     ciphersuite: &'static Ciphersuite,
@@ -212,7 +212,7 @@ fn read_test_vectors() {
     }
 }
 
-#[cfg(any(feature = "expose-test-vectors", test))]
+#[cfg(any(feature = "test-utils", test))]
 pub fn run_test_vector(test_vector: KeyScheduleTestVector) -> Result<(), KsTestVectorError> {
     use tls_codec::Serialize;
