+++ openmls/src/test_utils/mod.rs
@@ -7,7 +7,9 @@ use std::{
     io::{BufReader, Write},
 };
 
-pub fn write(file_name: &str, obj: impl Serialize) {
+pub mod test_framework;
+
+pub(crate) fn write(file_name: &str, obj: impl Serialize) {
     let mut file = match File::create(file_name) {
         Ok(f) => f,
         Err(_) => panic!("Couldn't open file {}.", file_name),
@@ -20,7 +22,7 @@ pub fn write(file_name: &str, obj: impl Serialize) {
     .expect("Error writing test vector file");
 }
 
-pub fn read<T: DeserializeOwned>(file_name: &str) -> T {
+pub(crate) fn read<T: DeserializeOwned>(file_name: &str) -> T {
     let file = match File::open(file_name) {
         Ok(f) => f,
         Err(_) => panic!("Couldn't open file {}.", file_name),
@@ -33,7 +35,7 @@ pub fn read<T: DeserializeOwned>(file_name: &str) -> T {
 }
 
 /// Convert `bytes` to a hex string.
-pub fn bytes_to_hex(bytes: &[u8]) -> String {
+pub(crate) fn bytes_to_hex(bytes: &[u8]) -> String {
     let mut hex = String::new();
     for &b in bytes {
         hex += &format!("{:02X}", b);
@@ -42,7 +44,7 @@ pub fn bytes_to_hex(bytes: &[u8]) -> String {
 }
 
 /// Convert a hex string to a byte vector.
-pub fn hex_to_bytes(hex: &str) -> Vec<u8> {
+pub(crate) fn hex_to_bytes(hex: &str) -> Vec<u8> {
     assert!(hex.len() % 2 == 0);
     let mut bytes = Vec::new();
     for i in 0..(hex.len() / 2) {
@@ -53,7 +55,7 @@ pub fn hex_to_bytes(hex: &str) -> Vec<u8> {
 
 /// Convert a hex string to a byte vector.
 /// If the input is `None`, this returns an empty vector.
-pub fn hex_to_bytes_option(hex: Option<String>) -> Vec<u8> {
+pub(crate) fn hex_to_bytes_option(hex: Option<String>) -> Vec<u8> {
     match hex {
         Some(s) => hex_to_bytes(&s),
         None => vec![],
