+++ openmls/src/tree/tests/mod.rs
@@ -1,10 +1,10 @@
 //! Tree unit tests.
 
-#[cfg(any(feature = "expose-test-vectors", test))]
+#[cfg(any(feature = "test-utils", test))]
 pub mod kat_encryption;
-#[cfg(any(feature = "expose-test-vectors", test))]
+#[cfg(any(feature = "test-utils", test))]
 pub mod kat_tree_kem;
-#[cfg(any(feature = "expose-test-vectors", test))]
+#[cfg(any(feature = "test-utils", test))]
 pub mod kat_treemath;
 
 #[cfg(test)]
